Update
action.js and style.css were added to the project file and linked to index.html.

I found that using <article>, <summary> and <details> semantic HTML tags will actually be more beneficial to completing this project than trying to assemble it via divs.

This update just got all the main elements of the page in place and working. Will work on stylizing and adding functionality. I plan on working on the functionality first, then styling after. The next thing that I want to work on would be using js to change the plus-icons to negative once the details are shown.