Big Update
I took out the JS completely, I was overthinking it and overcomplicating this challenge. I updated the index.html by adding the description meta tag, deleted the image tags (to add them on using the pseudo-selectors ::before and ::after in the style file). I also added the name="accordion" to the details tag to only allow 1 accordion tab to be open at a time. I began to further stylize the accordion to progress it closer to the original design.

I plan to finish the stylizing, and possibly utilize some JS for functionality. I may not need to do that , but that's for another time.